feat: liquidity provision provider refactor
